requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.7
anthropic>=0.40.0
feedparser>=6.0.10
feedgen>=1.0.0
apscheduler>=3.10.4
jinja2>=3.1.2
python-multipart>=0.0.6
aiofiles>=23.2.1
aiohttp>=3.9.0
python-dateutil>=2.8.2
pyyaml>=6.0.1
pytest>=7.4.3
pytest-asyncio>=0.21.1
httpx>=0.25.2
